+++ app/src/main/java/com/example/cattasticpos/domain/model/InventoryItem.kt
@@ -4,6 +4,6 @@ data class InventoryItem(
     val id: String,
     val itemName: String,
     val unit: String,
-    val currentStock: Int,
-    val reorderThreshold: Int
+    val currentStock: Double,
+    val reorderThreshold: Double
 )
